pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "reliafy"
version = "0.1.3"
description = "Reliafy CLI for reliability analysis, design optimization, and simulation."
readme = { file = "README.md", content-type = "text/markdown" }
authors = [{name = "Cesar Carrasco"}]
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "requests>=2.31.0",
  "numpy>=1.24.0",
  "matplotlib>=3.7.0",
  "pydantic>=2.0.0",
  "pydantic-core>=2.10.0",
  "typing_extensions>=4.7.0",
  "filedialpy>=0.1.0",
  "platformdirs>=3.0.0",
  "python-dotenv>=1.0.0",
  "PyJWT>=2.0.0",
  "PyYAML>=6.0",
]

[project.scripts]
reliafy = "reliafy.reliafy:app"

[project.urls]
Homepage = "https://reliafy.pages.dev"
Documentation = "https://reliafy.pages.dev"
Repository = "https://github.com/CarrascoCesar/Reliafy-CLI"

[tool.hatch.build.targets.wheel]
packages = ["reliafy"]
include = [
  "reliafy/examples/**",
]

[tool.hatch.build.targets.sdist]
include = [
  "reliafy/**",
  "reliafy/examples/**",
  "README.md",
]
